maint: Enable performance-enum-size check
Most enums fit into a uint8_t representation.

We have to name 'enum GjsError' and 'enum GjsJSError' separately from
their typedef due to https://github.com/llvm/llvm-project/issues/178114.

+++ gi/info.h
@@ -57,7 +57,7 @@ using BoolResult = mozilla::Result<mozilla::Ok, mozilla::Nothing>;
 
 namespace GI {
 
-enum class InfoTag : unsigned {
+enum class InfoTag : uint8_t {
     ARG,
     BASE,
     CALLABLE,
